+++ LeapEngine/CMakeLists.txt
@@ -25,11 +25,11 @@ add_library(LeapEngine STATIC
     "Components/Audio/AudioListener.cpp"
     "Components/RenderComponents/TerrainComponent.cpp"
     "Components/Physics/BoxCollider.cpp" 
-    "Utils/PhysicsSync.cpp" 
+    "Physics/PhysicsSync.cpp" 
     "Components/Physics/Rigidbody.cpp"
     "Components/Physics/SphereCollider.cpp"
     "Components/Physics/Collider.cpp" 
 	"Components/Physics/CapsuleCollider.cpp"
-	"Utils/Physics.cpp")
+	"Physics/Physics.cpp")
 
 set(LeapEngineIncludeDir "${CMAKE_CURRENT_SOURCE_DIR}" PARENT_SCOPE)
